# artifacts/api-server/src/routes/production.ts
import { Router } from "express";
import { db, productionRecordsTable, projectsTable, activityTable } from "@workspace/db";
import { eq, desc } from "drizzle-orm";
import {
  CreateProductionRecordBody,
  GetProductionRecordParams,
  DeleteProductionRecordParams,
} from "@workspace/api-zod";
import { requireRole, canAccessProject } from "../middlewares/auth";

const router = Router();

// GET /production — filter by project access
router.get("/", async (req, res) => {
  try {
    const projectIdRaw = req.query.projectId;
    const projectId = projectIdRaw ? String(projectIdRaw) : undefined;

    // If a specific projectId is requested, check access first
    if (projectId && !canAccessProject(req, projectId)) {
      res.status(403).json({ error: "Forbidden: no access to this project" });
      return;
    }

    const rows = await db
      .select({
        id: productionRecordsTable.id,
        projectId: productionRecordsTable.projectId,
        projectName: projectsTable.name,
        recordedAt: productionRecordsTable.recordedAt,
        productionKg: productionRecordsTable.productionKg,
        soldKg: productionRecordsTable.soldKg,
        sellingPricePerKg: productionRecordsTable.sellingPricePerKg,
        revenue: productionRecordsTable.revenue,
        notes: productionRecordsTable.notes,
        createdAt: productionRecordsTable.createdAt,
      })
      .from(productionRecordsTable)
      .innerJoin(projectsTable, eq(productionRecordsTable.projectId, projectsTable.id))
      .where(projectId ? eq(productionRecordsTable.projectId, projectId) : undefined)
      .orderBy(desc(productionRecordsTable.recordedAt));

    const formatted = rows.map((r) => ({
      ...r,
      recordedAt: r.recordedAt.toISOString(),
      createdAt: r.createdAt.toISOString(),
    }));

    if (req.canAccessAllProjects) {
      res.json(formatted);
    } else {
      res.json(formatted.filter((r) => canAccessProject(req, r.projectId)));
    }
  } catch (err) {
    req.log.error({ err }, "Failed to list production records");
    res.status(500).json({ error: "Internal server error" });
  }
});

// POST /production — admin, developer, or employee
router.post("/", requireRole("admin", "developer", "employee"), async (req, res) => {
  const parsed = CreateProductionRecordBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }

  const { projectId, recordedAt, productionKg, soldKg, sellingPricePerKg, notes } = parsed.data;

  if (!canAccessProject(req, projectId)) {
    res.status(403).json({ error: "Forbidden: no access to this project" });
    return;
  }

  const revenue = soldKg * sellingPricePerKg;

  try {
    const [project] = await db
      .select({ name: projectsTable.name })
      .from(projectsTable)
      .where(eq(projectsTable.id, projectId));
    if (!project) {
      res.status(404).json({ error: "Project not found" });
      return;
    }

    const [record] = await db
      .insert(productionRecordsTable)
      .values({
        projectId,
        recordedAt: new Date(recordedAt),
        productionKg,
        soldKg,
        sellingPricePerKg,
        revenue,
        notes: notes ?? null,
      })
      .returning();

    await db.insert(activityTable).values({
      type: "production_logged",
      description: `${soldKg} kg sold @ ₹${sellingPricePerKg}/kg for "${project.name}" — revenue ₹${revenue.toLocaleString("en-IN")}`,
      entityId: record.id,
      entityType: "production",
    });

    res.status(201).json({
      ...record,
      projectName: project.name,
      recordedAt: record.recordedAt.toISOString(),
      createdAt: record.createdAt.toISOString(),
    });
  } catch (err) {
    req.log.error({ err }, "Failed to create production record");
    res.status(500).json({ error: "Internal server error" });
  }
});

// GET /production/:id — check project access
router.get("/:id", async (req, res) => {
  const parsed = GetProductionRecordParams.safeParse({ id: req.params.id });
  if (!parsed.success) {
    res.status(400).json({ error: "Invalid id" });
    return;
  }

  try {
    const [row] = await db
      .select({
        id: productionRecordsTable.id,
        projectId: productionRecordsTable.projectId,
        projectName: projectsTable.name,
        recordedAt: productionRecordsTable.recordedAt,
        productionKg: productionRecordsTable.productionKg,
        soldKg: productionRecordsTable.soldKg,
        sellingPricePerKg: productionRecordsTable.sellingPricePerKg,
        revenue: productionRecordsTable.revenue,
        notes: productionRecordsTable.notes,
        createdAt: productionRecordsTable.createdAt,
      })
      .from(productionRecordsTable)
      .innerJoin(projectsTable, eq(productionRecordsTable.projectId, projectsTable.id))
      .where(eq(productionRecordsTable.id, parsed.data.id));

    if (!row) {
      res.status(404).json({ error: "Not found" });
      return;
    }
    if (!canAccessProject(req, row.projectId)) {
      res.status(403).json({ error: "Forbidden: no access to this project" });
      return;
    }

    res.json({
      ...row,
      recordedAt: row.recordedAt.toISOString(),
      createdAt: row.createdAt.toISOString(),
    });
  } catch (err) {
    req.log.error({ err }, "Failed to get production record");
    res.status(500).json({ error: "Internal server error" });
  }
});

// DELETE /production/:id — admin or developer only
router.delete("/:id", requireRole("admin", "developer"), async (req, res) => {
  const parsed = DeleteProductionRecordParams.safeParse({ id: req.params.id });
  if (!parsed.success) {
    res.status(400).json({ error: "Invalid id" });
    return;
  }

  try {
    await db
      .delete(productionRecordsTable)
      .where(eq(productionRecordsTable.id, parsed.data.id));
    res.status(204).send();
  } catch (err) {
    req.log.error({ err }, "Failed to delete production record");
    res.status(500).json({ error: "Internal server error" });
  }
});

export default router;