# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Rubber Plantation Partnership API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /partners:
    get:
      operationId: listPartners
      tags: [partners]
      summary: List all partners
      responses:
        "200":
          description: List of partners
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Partner"
    post:
      operationId: createPartner
      tags: [partners]
      summary: Register a new partner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PartnerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Partner"
  /partners/{id}:
    get:
      operationId: getPartner
      tags: [partners]
      summary: Get a partner by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Partner details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Partner"
        "404":
          description: Not found
    patch:
      operationId: updatePartner
      tags: [partners]
      summary: Update a partner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PartnerUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Partner"
  /agreements:
    get:
      operationId: listAgreements
      tags: [agreements]
      summary: List all partnership agreements
      responses:
        "200":
          description: List of agreements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Agreement"
    post:
      operationId: createAgreement
      tags: [agreements]
      summary: Create a new partnership agreement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgreementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agreement"
  /agreements/{id}:
    get:
      operationId: getAgreement
      tags: [agreements]
      summary: Get an agreement by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Agreement details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agreement"
        "404":
          description: Not found
    patch:
      operationId: updateAgreement
      tags: [agreements]
      summary: Update an agreement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgreementUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agreement"
...
  /production:
    get:
      operationId: listProductionRecords
      tags: [production]
      summary: List all rubber production records
      parameters:
        - name: projectId
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of production records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductionRecord"
    post:
      operationId: createProductionRecord
      tags: [production]
      summary: Log a new rubber production and sale record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductionRecord"
  /production/{id}:
    get:
      operationId: getProductionRecord
      tags: [production]
      summary: Get a production record by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Production record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductionRecord"
        "404":
          description: Not found
    delete:
      operationId: deleteProductionRecord
      tags: [production]
      summary: Delete a production record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Partner:
      type: object
      required: [id, name, role, createdAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        role:
          type: string
          enum: [project_developer, landowner, investor]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        aadhaarLast4:
          type: ["string", "null"]
        clerkUserId:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
    PartnerInput:
      type: object
      required: [name, role]
      properties:
        name:
          type: string
        role:
          type: string
          enum: [project_developer, landowner, investor]
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        aadhaarLast4:
          type: string
        clerkUserId:
          type: string
        notes:
          type: string
    PartnerUpdate:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
          enum: [project_developer, landowner, investor]
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        aadhaarLast4:
          type: string
        notes:
          type: string
    Agreement:
      type: object
      required: [id, projectId, projectName, landOwnerId, landOwnerName, projectDeveloperId, projectDeveloperName, executionDate, executionPlace, termYears, landArea, landAreaUnit, landNotionalValue, landValuePerUnit, landContributionAdjustment, yearlyEscalation, status, createdAt]
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        projectName:
          type: string
        landOwnerId:
          type: string
          format: uuid
        landOwnerName:
          type: string
        projectDeveloperId:
          type: string
          format: uuid
        projectDeveloperName:
          type: string
        executionDate:
          type: string
        executionPlace:
          type: string
        termYears:
          type: integer
        landArea:
          type: number
        landAreaUnit:
          type: string
        landNotionalValue:
          type: number
        landValuePerUnit:
          type: number
        landContributionAdjustment:
          type: number
        yearlyEscalation:
          type: number
        ownershipShareLandowner:
          type: ["number", "null"]
        ownershipShareDeveloper:
          type: ["number", "null"]
        revenueModel:
          type: string
          enum: [contribution, fifty_percent_revenue]
        status:
          type: string
          enum: [draft, active, matured, terminated]
        northBoundary:
          type: ["string", "null"]
        southBoundary:
          type: ["string", "null"]
        eastBoundary:
          type: ["string", "null"]
        westBoundary:
          type: ["string", "null"]
        gpsLat:
          type: ["number", "null"]
        gpsLng:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
    AgreementInput:
      type: object
      required: [projectId, landOwnerId, projectDeveloperId, executionDate, executionPlace, termYears, landArea, landAreaUnit, landNotionalValue, landValuePerUnit, landContributionAdjustment, yearlyEscalation]
      properties:
        projectId:
          type: string
          format: uuid
        landOwnerId:
          type: string
          format: uuid
        projectDeveloperId:
          type: string
          format: uuid
        executionDate:
          type: string
        executionPlace:
          type: string
        termYears:
          type: integer
        landArea:
          type: number
        landAreaUnit:
          type: string
        landNotionalValue:
          type: number
        landValuePerUnit:
          type: number
        landContributionAdjustment:
          type: number
        yearlyEscalation:
          type: number
        revenueModel:
          type: string
          enum: [contribution, fifty_percent_revenue]
        northBoundary:
          type: string
        southBoundary:
          type: string
        eastBoundary:
          type: string
        westBoundary:
          type: string
        gpsLat:
          type: number
        gpsLng:
          type: number
        notes:
          type: string
    AgreementUpdate:
      type: object
      properties:
        executionDate:
          type: string
        executionPlace:
          type: string
        termYears:
          type: integer
        landArea:
          type: number
        landAreaUnit:
          type: string
        landNotionalValue:
          type: number
        landValuePerUnit:
          type: number
        landContributionAdjustment:
          type: number
        yearlyEscalation:
          type: number
        ownershipShareLandowner:
          type: number
        ownershipShareDeveloper:
          type: number
        revenueModel:
          type: string
          enum: [contribution, fifty_percent_revenue]
        status:
          type: string
          enum: [draft, active, matured, terminated]
        northBoundary:
          type: string
        southBoundary:
          type: string
        eastBoundary:
          type: string
        westBoundary:
          type: string
        gpsLat:
          type: number
        gpsLng:
          type: number
        notes:
          type: string
...
    ProductionRecord:
      type: object
      required: [id, projectId, projectName, recordedAt, productionKg, soldKg, sellingPricePerKg, revenue, createdAt]
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        projectName:
          type: string
        recordedAt:
          type: string
        productionKg:
          type: number
        soldKg:
          type: number
        sellingPricePerKg:
          type: number
        revenue:
          type: number
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    ProductionInput:
      type: object
      required: [projectId, recordedAt, productionKg, soldKg, sellingPricePerKg]
      properties:
        projectId:
          type: string
          format: uuid
        recordedAt:
          type: string
          description: ISO datetime of when production/sale happened
        productionKg:
          type: number
          description: Total rubber produced in kg
        soldKg:
          type: number
          description: Quantity sold in kg
        sellingPricePerKg:
          type: number
          description: Selling price per kg in INR
        notes:
          type: string
...