# lib/api-zod/src/generated/types/agreement.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rubber Plantation Partnership API
 * OpenAPI spec version: 0.1.0
 */
import type { AgreementRevenueModel } from "./agreementRevenueModel";
import type { AgreementStatus } from "./agreementStatus";

export interface Agreement {
  id: string;
  projectId: string;
  projectName: string;
  landOwnerId: string;
  landOwnerName: string;
  projectDeveloperId: string;
  projectDeveloperName: string;
  executionDate: string;
  executionPlace: string;
  termYears: number;
  landArea: number;
  landAreaUnit: string;
  landNotionalValue: number;
  landValuePerUnit: number;
  landContributionAdjustment: number;
  yearlyEscalation: number;
  /** @nullable */
  ownershipShareLandowner?: number | null;
  /** @nullable */
  ownershipShareDeveloper?: number | null;
  revenueModel?: AgreementRevenueModel;
  status: AgreementStatus;
  /** @nullable */
  northBoundary?: string | null;
  /** @nullable */
  southBoundary?: string | null;
  /** @nullable */
  eastBoundary?: string | null;
  /** @nullable */
  westBoundary?: string | null;
  /** @nullable */
  gpsLat?: number | null;
  /** @nullable */
  gpsLng?: number | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}