# lib/api-zod/src/generated/types/agreementInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rubber Plantation Partnership API
 * OpenAPI spec version: 0.1.0
 */
import type { AgreementInputRevenueModel } from "./agreementInputRevenueModel";

export interface AgreementInput {
  projectId: string;
  landOwnerId: string;
  projectDeveloperId: string;
  executionDate: string;
  executionPlace: string;
  termYears: number;
  landArea: number;
  landAreaUnit: string;
  landNotionalValue: number;
  landValuePerUnit: number;
  landContributionAdjustment: number;
  yearlyEscalation: number;
  revenueModel?: AgreementInputRevenueModel;
  northBoundary?: string;
  southBoundary?: string;
  eastBoundary?: string;
  westBoundary?: string;
  gpsLat?: number;
  gpsLng?: number;
  notes?: string;
}