# lib/api-zod/src/generated/types/partner.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rubber Plantation Partnership API
 * OpenAPI spec version: 0.1.0
 */
import type { PartnerRole } from "./partnerRole";

export interface Partner {
  id: string;
  name: string;
  role: PartnerRole;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  aadhaarLast4?: string | null;
  /** @nullable */
  clerkUserId?: string | null;
  /** @nullable */
  notes?: string | null;
  isActive?: boolean;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}